requirements.txt
openai>=1.75.0
pandas>=2.2.3
openpyxl>=3.1.5
Pillow>=11.2.1